+++ api/User.php
@@ -102,7 +102,13 @@ class User {
 //    $BoardDB = Board::connect();
     $mysqli = User::connect();
     $result0 = $mysqli->query("delete from Item where user_id = \"$this->id\"");
+    if(!$result0){
+          return false;
+    }
     $result1 = $mysqli->query("delete from Board where user_id = \"$this->id\"");
+    if(!$result1){
+          return false;
+    }
     $result = $mysqli->query("delete FROM User where id = \"$this->id\"");
     return $result;
   }
